refactor: split the main app and site layout (#740)
* refactor: split the main app and site layout

* refactor: finish merge conflict

* fix: remove obsolete check

* refactor: move all layout files into same folder

* perf: scope the style

* perf: improve site route matching

* fix: finish merge conflict
