+++ apps/ui/src/components/Layout/Site.vue
@@ -1,0 +1,5 @@
+<template>
+  <SiteTopnav />
+  <router-view />
+  <SiteFooter />
+</template>
