+++ apps/ui/src/views/Site.vue
@@ -1,7 +1,3 @@
 <template>
-  <div class="!mt-0 -mb-6">
-    <SiteTopnav />
-    <router-view class="min-h-[calc(100vh-320px)]" />
-    <SiteFooter />
-  </div>
+  <router-view class="min-h-[calc(100vh-320px)]" />
 </template>
